+++ src/apscheduler/triggers/cron/__init__.py
@@ -94,7 +94,6 @@ class CronTrigger(Trigger):
                 self.second,
             ]
         )
-        self._last_fire_time: datetime | None = None
 
     def _set_fields(self, values: Sequence[int | str | None]) -> None:
         self._fields = []
@@ -112,7 +111,14 @@ class CronTrigger(Trigger):
             self._fields.append(field)
 
     @classmethod
-    def from_crontab(cls, expr: str, timezone: str | tzinfo = "local") -> CronTrigger:
+    def from_crontab(
+        cls,
+        expr: str,
+        *,
+        start_time: datetime | None = None,
+        end_time: datetime | None = None,
+        timezone: str | tzinfo = "local",
+    ) -> CronTrigger:
         """
         Create a :class:`~CronTrigger` from a standard crontab expression.
 
@@ -120,6 +126,9 @@ class CronTrigger(Trigger):
         accepted here.
 
         :param expr: minute, hour, day of month, month, day of week
+        :param start_time: earliest possible date/time to trigger on (defaults to current
+            time)
+        :param end_time: latest possible date/time to trigger on
         :param timezone: time zone to use for the date/time calculations
             (defaults to local timezone if omitted)
 
@@ -134,6 +143,8 @@ class CronTrigger(Trigger):
             day=values[2],
             month=values[3],
             day_of_week=values[4],
+            start_time=start_time or datetime.now(),
+            end_time=end_time,
             timezone=timezone,
         )